+++ R/chk-is.R
@@ -48,6 +48,86 @@ chk_is <- function(x, class, x_name = NULL) {
 #' vld_is(NULL, "numeric")
 vld_is <- function(x, class) inherits(x, class)
 
+#' @describeIn chk_is Check Inherits from S3 Class
+#'
+#' Checks inherits from S3 class using
+#'
+#' \code{!isS4(x) && \link{inherits}(x, class)}
+#'
+#' Class should be a string.
+#'
+#' @export
+#'
+#' @examples
+#'
+#' # chk_s3_class
+#' chk_s3_class(1, "numeric")
+#' try(chk_s3_class(getClass("MethodDefinition"), "classRepresentation"))
+chk_s3_class <- function(x, class, x_name = NULL) {
+  if (vld_s3_class(x, class)) {
+    return(invisible())
+  }
+  if (is.null(x_name)) x_name <- deparse_backtick(substitute(x))
+  abort_chk(x_name, " must inherit from S3 class '", class, "'.")
+}
+
+#' @describeIn chk_is Validate Inherits from S3 Class
+#'
+#' Validates inherits from S3 class
+#'
+#' \code{!isS4(x) && \link{inherits}(x, class)}
+#'
+#' Class should be a string.
+#'
+#' @export
+#'
+#' @examples
+#'
+#' # vld_s3_class
+#' vld_s3_class(numeric(0), "numeric")
+#' vld_s3_class(getClass("MethodDefinition"), "classRepresentation")
+vld_s3_class <- function(x, class) !isS4(x) && inherits(x, class)
+
+#' @describeIn chk_is Check Inherits from S4 Class
+#'
+#' Checks inherits from S4 class using
+#'
+#' \code{isS4(x) && \link{inherits}(x, class)}
+#'
+#' Class should be a string.
+#'
+#' @export
+#'
+#' @examples
+#'
+#' # chk_s4_class
+#' try(chk_s4_class(1, "numeric"))
+#' chk_s4_class(getClass("MethodDefinition"), "classRepresentation")
+chk_s4_class <- function(x, class, x_name = NULL) {
+  if (vld_s4_class(x, class)) {
+    return(invisible())
+  }
+  if (is.null(x_name)) x_name <- deparse_backtick(substitute(x))
+  abort_chk(x_name, " must inherit from S4 class '", class, "'.")
+}
+
+#' @describeIn chk_is Validate Inherits from S4 Class
+#'
+#' Validates inherits from S4 class
+#'
+#' \code{isS4(x) && \link{inherits}(x, class)}
+#'
+#' Class should be a string.
+#'
+#' @export
+#'
+#' @examples
+#'
+#' # vld_s4_class
+#' vld_s4_class(numeric(0), "numeric")
+#' vld_s4_class(getClass("MethodDefinition"), "classRepresentation")
+vld_s4_class <- function(x, class) isS4(x) && inherits(x, class)
+
 #' @describeIn chk_is Check Whole Numeric
 #'
 #' Checks if integer vector or double equivalent using \code{vld_whole_numeric()}.